DDOS-218 - added plain view controller example
