+++ BottomSheet/Examples/Plain/Base.lproj/Plain.storyboard
@@ -8,14 +8,21 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--View Controller-->
+        <!--Plain View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="ViewController" customModule="BottomSheet" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="BYZ-38-t0r" customClass="PlainViewController" customModule="BottomSheet" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="320" height="568"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="w1P-Ab-sMK">
+                                <rect key="frame" x="133" y="40" width="54" height="30"/>
+                                <state key="normal" title="Dismiss"/>
+                                <connections>
+                                    <action selector="dismiss" destination="BYZ-38-t0r" eventType="touchUpInside" id="svy-br-fC1"/>
+                                </connections>
+                            </button>
                             <containerView opaque="NO" contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Nby-6d-aeB">
                                 <rect key="frame" x="0.0" y="0.0" width="320" height="568"/>
                                 <connections>
@@ -27,8 +34,10 @@
                         <constraints>
                             <constraint firstItem="Nby-6d-aeB" firstAttribute="leading" secondItem="6Tk-OE-BBY" secondAttribute="leading" id="ETV-xo-1cE"/>
                             <constraint firstItem="Nby-6d-aeB" firstAttribute="height" secondItem="6Tk-OE-BBY" secondAttribute="height" id="G2B-gm-yGO"/>
+                            <constraint firstItem="w1P-Ab-sMK" firstAttribute="top" secondItem="6Tk-OE-BBY" secondAttribute="top" constant="40" id="M5K-ce-6Ef"/>
                             <constraint firstItem="6Tk-OE-BBY" firstAttribute="trailing" secondItem="Nby-6d-aeB" secondAttribute="trailing" id="Mic-uy-Mjx"/>
                             <constraint firstItem="Nby-6d-aeB" firstAttribute="bottom" secondItem="6Tk-OE-BBY" secondAttribute="bottom" id="yIo-Rd-KSj"/>
+                            <constraint firstItem="w1P-Ab-sMK" firstAttribute="centerX" secondItem="8bC-Xf-vdC" secondAttribute="centerX" id="ziv-Nf-WmM"/>
                         </constraints>
                         <viewLayoutGuide key="safeArea" id="6Tk-OE-BBY"/>
                     </view>
