+++ BottomSheet/Main.storyboard
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="16096" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="EnS-dm-mQd">
+    <device id="retina4_0" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--View Controller-->
+        <scene sceneID="LnY-zm-fbv">
+            <objects>
+                <viewController id="EnS-dm-mQd" customClass="ViewController" customModule="BottomSheet" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="bLf-G6-oky">
+                        <rect key="frame" x="0.0" y="0.0" width="320" height="568"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="UV2-rf-GeO">
+                                <rect key="frame" x="89" y="40" width="142" height="30"/>
+                                <state key="normal" title="Plain View Controller"/>
+                                <connections>
+                                    <action selector="plainViewController" destination="EnS-dm-mQd" eventType="touchUpInside" id="zlO-HX-OMV"/>
+                                </connections>
+                            </button>
+                        </subviews>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
+                        <constraints>
+                            <constraint firstItem="UV2-rf-GeO" firstAttribute="centerX" secondItem="bLf-G6-oky" secondAttribute="centerX" id="fqc-0G-iJc"/>
+                            <constraint firstItem="UV2-rf-GeO" firstAttribute="top" secondItem="kWs-0D-Gb7" secondAttribute="top" constant="40" id="odU-CC-z1c"/>
+                        </constraints>
+                        <viewLayoutGuide key="safeArea" id="kWs-0D-Gb7"/>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="3Bj-ld-ga8" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="51" y="120"/>
+        </scene>
+    </scenes>
+</document>
