+++ BottomSheet/ViewController.swift
@@ -8,22 +8,14 @@
 import UIKit
 
 class ViewController: UIViewController {
-    @IBOutlet weak var animator: BottomSheetAnimator!
-    
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        switch segue.destination {
-        case let viewController as TableViewController:
-            viewController.onDidScroll = { [weak self] scrollView in
-                self?.animator.scrollViewDidScroll(scrollView)
-            }
-            viewController.onViewDidLoad = { [weak self] in
-                guard let strongSelf = self else { return }
-                strongSelf.animator.animatableScrollView = viewController.tableView
-                strongSelf.animator.availablePositions = [.bottom, .middle, .custom(1)]
-                strongSelf.animator.currentPosition = .bottom
-            }
-        default:
-            super.prepare(for: segue, sender: sender)
-        }
+    @IBAction func plainViewController() {
+        guard let vc = UIStoryboard(name: "Plain", bundle: nil).instantiateInitialViewController() else { return }
+        present(vc, animated: true)
+    }
+}
+
+extension UIViewController {
+    @IBAction func dismiss() {
+        dismiss(animated: true)
     }
 }
